Refactoring api call logic
