--- api.ts
@@ -1,2 +1,0 @@
-export const backendApiUrl = "https://hacku-kosen-2022-backend-id432d3htq-an.a.run.app";
-export const getBackendApiKey = () => process.env.NEXT_PUBLIC_BACKEND_API_KEY;
